Task 0

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -8,7 +8,7 @@
 size_t print_dlistint(const dlistint_t *h)
 {
 	const dlistint_t *current;
-	int cpt = 0;
+	size_t cpt = 0;
 
 	current = h;
 	while (current)
